requirements.txt
langchain==0.1.9
langchain_openai==0.0.8
langchain_pinecone==0.0.3
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.28.2
